pyproject.toml
[project]
name = "ruuvi-listener-tools"
version = "0.1.0"
description = "Tools for developing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
]
